seperating server side from client side (attempt to store token on local)
