+++ src/app/loggedin/page.js
@@ -1,98 +1,107 @@
 'use client';
 import { useSearchParams } from "next/navigation";
+import getToken from "../token/token";
+import generateRecSong from "../api/token"
+
 import React from 'react';
+
+
+
 
 let client_id = process.env.SPOTIFY_CLIENT_ID;
 let client_secret = process.env.SPOTIFY_CLIENT_SECRET;
-let redirect_uri = 'http://localhost:3000/loggedin';
+let redirect_uri = 'http://localhost:3000/api';
 
 export default async function Home({ params, searchParams }) {
-  const [recommendedSong, setRecommendedSong] = React.useState({});
-  const code = searchParams?.code;
+//   const [recommendedSong, setRecommendedSong] = React.useState({});
+//   const code = searchParams?.code;
 
-  const url = 'https://accounts.spotify.com/api/token';
-  const body = new URLSearchParams({
-    grant_type: 'authorization_code',
-    code: code,
-    redirect_uri: redirect_uri,
-  });
+//   const url = 'https://accounts.spotify.com/api/token';
+//   const body = new URLSearchParams({
+//     grant_type: 'authorization_code',
+//     code: code,
+//     redirect_uri: redirect_uri,
+//   });
   
-  const headers = {
-    'Content-Type': 'application/x-www-form-urlencoded',
-    'Authorization': 'Basic ' + btoa(client_id + ':' + client_secret),
-  };
+//   const headers = {
+//     'Content-Type': 'application/x-www-form-urlencoded',
+//     'Authorization': 'Basic ' + btoa(client_id + ':' + client_secret),
+//   };
 
-  const response = await fetch(url, {
-    method: 'POST',
-    headers: headers,
-    body: body,
-  });
+//   const response = await fetch(url, {
+//     method: 'POST',
+//     headers: headers,
+//     body: body,
+//   });
 
-  if (!response.ok) {
-    const errorData = await response.json();
-    console.error('Error fetching access token:', errorData);
-    return <div>Error fetching data</div>;
-  }
+//   if (!response.ok) {
+//     const errorData = await response.json();
+//     console.error('Error fetching access token:', errorData);
+//     return <div>Error fetching data</div>;
+//   }
 
-  const tokenData = await response.json();
-  const auth_header = {
-    Authorization: 'Bearer ' + tokenData.access_token,
-  };
+//   const tokenData = await response.json();
+//   const auth_header = {
+//     Authorization: 'Bearer ' + tokenData.access_token,
+//   };
 
-  // Fetch top artists
-  const artistsResponse = await fetch("https://api.spotify.com/v1/me/top/artists?limit=4", {
-    headers: auth_header,
-  });
+//   // Fetch top artists
+//   const artistsResponse = await fetch("https://api.spotify.com/v1/me/top/artists?limit=4", {
+//     headers: auth_header,
+//   });
 
-  if (!artistsResponse.ok) {
-    const errorData = await artistsResponse.json();
-    console.error('Error fetching top artists:', errorData);
-    return <div>Error fetching data</div>;
-  }
+//   if (!artistsResponse.ok) {
+//     const errorData = await artistsResponse.json();
+//     console.error('Error fetching top artists:', errorData);
+//     return <div>Error fetching data</div>;
+//   }
 
-  const artists = await artistsResponse.json();
+//   const artists = await artistsResponse.json();
 
-  // Fetch top songs
-  const topSongsResponse = await fetch("https://api.spotify.com/v1/me/top/tracks?limit=4", {
-    headers: auth_header,
-  });
+//   // Fetch top songs
+//   const topSongsResponse = await fetch("https://api.spotify.com/v1/me/top/tracks?limit=4", {
+//     headers: auth_header,
+//   });
 
-  if (!topSongsResponse.ok) {
-    const errorData = await topSongsResponse.json();
-    console.error('Error fetching top songs:', errorData);
-    return <div>Error fetching data</div>;
-  }
+//   if (!topSongsResponse.ok) {
+//     const errorData = await topSongsResponse.json();
+//     console.error('Error fetching top songs:', errorData);
+//     return <div>Error fetching data</div>;
+//   }
 
-  const topSongs = await topSongsResponse.json();
-  const songIds = topSongs.items.map(song => song.id);
+//   const topSongs = await topSongsResponse.json();
+//   const songIds = topSongs.items.map(song => song.id);
 
-  // Fetch recommended songs based on top 5 songs
-const recommendedSongResponse = await fetch(
-  `https://api.spotify.com/v1/recommendations?limit=10&seed_tracks=${songIds.join()}`,
-  {
-    headers: auth_header,
-  }
-);
-const recommendedSongs = await recommendedSongResponse.json();
+//   // Fetch recommended songs based on top 5 songs
+// const recommendedSongResponse = await fetch(
+//   `https://api.spotify.com/v1/recommendations?limit=10&seed_tracks=${songIds.join()}`,
+//   {
+//     headers: auth_header,
+//   }
+// );
+// const recommendedSongs = await recommendedSongResponse.json();
 
 
 
-// Function to obtain a random song from recommended songs
-const getRandomSong = () => {
-  const randomIndex = Math.floor(Math.random() * recommendedSongs.tracks.length);
-  return recommendedSongs.tracks[randomIndex];
-}
-const recomSong = getRandomSong();
-setRecommendedSong(recomSong);
-console.log("This is a new thing", recomSong);
-console.log("This is a new thing", recomSong.name);
-console.log("This is should match above", recommendedSong.name);
+// // Function to obtain a random song from recommended songs
+// const getRandomSong = () => {
+//   const randomIndex = Math.floor(Math.random() * recommendedSongs.tracks.length);
+//   return recommendedSongs.tracks[randomIndex];
+// }
+// const recomSong = getRandomSong();
+// setRecommendedSong(recomSong);
+// console.log("This is a new thing", recomSong);
+// console.log("This is a new thing", recomSong.name);
+// console.log("This is should match above", recommendedSong.name);
 
-// Method to obtain user's display name
-const userResponse = await fetch("https://api.spotify.com/v1/me", {
-  headers: auth_header,
-});
-const user_data = await userResponse.json();
+// // Method to obtain user's display name
+// const userResponse = await fetch("https://api.spotify.com/v1/me", {
+//   headers: auth_header,
+// });
+// const user_data = await userResponse.json();
+
+console.log(getToken());
+
 
 return (
   <div className="relative flex flex-col items-center justify-center min-h-screen bg-cover bg-center text-white">
@@ -147,12 +156,6 @@ return (
           {`Recommended Song: ${recommendedSong.name}`}
         </a>
       )}
-      <button onClick={() => {
-        const randomSong = getRandomSong();
-        alert(`Random Recommended Song: ${randomSong.name}`);
-      }} className="text-xl font-bold bg-gray-700 p-3 rounded-lg shadow-lg mt-3">
-        Get New Recommended Song
-      </button>
     </div>
   </div>
 );
